Shoal integration for the bug fix below. Shoal fix by Joe, reviewed by yours truly. Jane ok'ed the pom change, and Sathyan said we could integrate it if I got it in about 8 minutes ago.
Fix for glassfish 16422.  This fixes bug that only occurs when one deletes and recreates a cluster with same name.
Workaround would be to restart DAS everytime a cluster is deleted and one plans on recreating the deleted cluster.

This is a low risk fix to ensure that distributed state cache is cleared of stale data when a cluster is deleted.
The DAS only leaves a cluster when the cluster is deleted (or DAS is shutting down.



git-svn-id: https://svn.java.net/svn/glassfish~svn/branches/3.1.1@47807 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5