Adding @NotNull for IiopService and JmsService and defaultValue for JmsService:
1) defaultValue="EMBEDDED" to getType() on JmsService
2) Add @NotNull to getIiopService() on Config bean
3) Add @NotNull to getJmsService() on Config bean

The newly added template web-domain.xml for web profile does not contain iiop-service and  jms-service under server-config.  In order to allow the user to be able to get/set and create these elements, using the commands create/delete/list-iiop-listener and create/delete/list-jms-host, @NotNull needs to be added for IiopService and JmsService to prevent an NPE.  The type attribute is also required for JmsService, so a defaultValue of 'EMBEDDED' is added.

git-svn-id: https://svn.java.net/svn/glassfish~svn/branches/3.1.2@48974 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5