Overlooked part of the fix to 13092305
When the administrator sets up secure admin to use a username and pw alias (instead of SSL certs) for interprocess authentication, the default aliases might not even be present in the keystore.  In such a case the enable-secure-admin logic was still validating the default value which failed.

These changes bypass the validation if the user does not indicate a non-default value for the alias(es).

Earlier approval: Joe
Tests: QL, manual tests for this situation

git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@51885 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5