Checking in on behalf of Kshitiz who did this as part of LB, DNS plugin related implementation along with minor changes.
+Adding support for bi-directional association/dissociation.
For each service-reference, Orchestrator will now call associate/dissociateServices
on both serviceConsumer and serviceProvider.
+Making sure that the order in which the associate/dissociate calls are made
is in accordance with the definition of <service-reference> in glassfish-services.xml,
the list of service-references provided by each plugin. So, if a plugin has provided
two service-references, associate/dissociate calls will be made in the same
order in which the service-references were provided by the plugin.


Tests Passed : QL (Web, Classic), PaaS tests


git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@52594 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5