Fix for 18773.
In certain cases if the DAS does not shut down in response to "stop-domain --force"  the local command logic skipped the code to kill the process because a lower-level exception was thrown but not caught.

These changes catch that exception and, if the user has specified "--kill" uses the OS kill feature.



git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@54370 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5