GLASSFISH-18880: Fail to open OSGi console from GF admin console or URL
This is continuation of my previous check in. My previous checkin caused a regression in
embedded ejb test suite as I had made the same class both a Startup service as well as an activator.
Since in some embedded distribution, we don't have OSGi classes, there the startup service failed
to load due to NCDFE. In this check in, I have separated them.

git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@56547 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5